requirements.txt
imageio
pillow-avif-plugin
rawpy
